# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Maintenance Report Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type ReportStatus = (typeof ReportStatus)[keyof typeof ReportStatus];

export const ReportStatus = {
  pending: "pending",
  in_progress: "in_progress",
  resolved: "resolved",
  cancelled: "cancelled",
} as const;

export type ReportPriority =
  (typeof ReportPriority)[keyof typeof ReportPriority];

export const ReportPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  reporter: "reporter",
  technician: "technician",
  admin: "admin",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  createdAt: string;
}

export interface CreateUser {
  name: string;
  email: string;
  role: UserRole;
}

export interface Equipment {
  id: number;
  name: string;
  location: string;
  description?: string;
  createdAt: string;
}

export interface CreateEquipment {
  name: string;
  location: string;
  description?: string;
}

export interface Report {
  id: number;
  title: string;
  description: string;
  status: ReportStatus;
  priority: ReportPriority;
  equipmentId: number;
  equipment?: Equipment;
  reportedById: number;
  reportedBy?: User;
  receivedById?: number;
  receivedBy?: User;
  assignedToId?: number;
  assignedTo?: User;
  resolvedById?: number;
  resolvedBy?: User;
  notes?: string;
  createdAt: string;
  updatedAt: string;
  resolvedAt?: string;
  /** Resolution time in minutes */
  resolutionTime?: number;
}

export interface CreateReport {
  title: string;
  description: string;
  priority: ReportPriority;
  equipmentId: number;
  reportedById: number;
  receivedById?: number;
}

export interface UpdateReport {
  status?: ReportStatus;
  assignedToId?: number;
  receivedById?: number;
  resolvedById?: number;
  notes?: string;
  title?: string;
  description?: string;
  priority?: ReportPriority;
}

export interface Stats {
  total: number;
  pending: number;
  inProgress: number;
  resolved: number;
  cancelled: number;
  /** Average resolution time in minutes */
  avgResolutionTime: number;
}

export type ListReportsParams = {
  status?: ReportStatus;
  assignedTo?: number;
  equipmentId?: number;
};