# lib/api-zod/src/generated/types/createEquipment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Maintenance Report Management API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateEquipment {
  name: string;
  location: string;
  description?: string;
}