# lib/api-zod/src/generated/types/createReport.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Maintenance Report Management API
 * OpenAPI spec version: 0.1.0
 */
import type { ReportPriority } from "./reportPriority";

export interface CreateReport {
  title: string;
  description: string;
  priority: ReportPriority;
  equipmentId: number;
  reportedById: number;
  receivedById?: number;
}