# lib/api-zod/src/generated/types/createUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Maintenance Report Management API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";

export interface CreateUser {
  name: string;
  email: string;
  role: UserRole;
}