# lib/api-zod/src/generated/types/report.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Maintenance Report Management API
 * OpenAPI spec version: 0.1.0
 */
import type { Equipment } from "./equipment";
import type { ReportPriority } from "./reportPriority";
import type { ReportStatus } from "./reportStatus";
import type { User } from "./user";

export interface Report {
  id: number;
  title: string;
  description: string;
  status: ReportStatus;
  priority: ReportPriority;
  equipmentId: number;
  equipment?: Equipment;
  reportedById: number;
  reportedBy?: User;
  receivedById?: number;
  receivedBy?: User;
  assignedToId?: number;
  assignedTo?: User;
  resolvedById?: number;
  resolvedBy?: User;
  notes?: string;
  createdAt: Date;
  updatedAt: Date;
  resolvedAt?: Date;
  /** Resolution time in minutes */
  resolutionTime?: number;
}