# lib/api-zod/src/generated/types/reportPriority.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Maintenance Report Management API
 * OpenAPI spec version: 0.1.0
 */

export type ReportPriority =
  (typeof ReportPriority)[keyof typeof ReportPriority];

export const ReportPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;