# lib/api-zod/src/generated/types/reportStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Maintenance Report Management API
 * OpenAPI spec version: 0.1.0
 */

export type ReportStatus = (typeof ReportStatus)[keyof typeof ReportStatus];

export const ReportStatus = {
  pending: "pending",
  in_progress: "in_progress",
  resolved: "resolved",
  cancelled: "cancelled",
} as const;