# lib/api-zod/src/generated/types/stats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Maintenance Report Management API
 * OpenAPI spec version: 0.1.0
 */

export interface Stats {
  total: number;
  pending: number;
  inProgress: number;
  resolved: number;
  cancelled: number;
  /** Average resolution time in minutes */
  avgResolutionTime: number;
}