# lib/api-zod/src/generated/types/updateReport.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Maintenance Report Management API
 * OpenAPI spec version: 0.1.0
 */
import type { ReportPriority } from "./reportPriority";
import type { ReportStatus } from "./reportStatus";

export interface UpdateReport {
  status?: ReportStatus;
  assignedToId?: number;
  receivedById?: number;
  resolvedById?: number;
  notes?: string;
  title?: string;
  description?: string;
  priority?: ReportPriority;
}